Minor debugging message bugfixes.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3811 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_net.c
@@ -700,7 +700,8 @@ mx_network_wait_for_message_id( MX_RECORD *server_record,
 #if 1
 		MX_DEBUG(-2,
 		("%s: MARKER 1: received_message_id = %#lx, message_id = %#lx",
-		 fname, received_message_id, message_id));
+		 fname, (unsigned long) received_message_id,
+		 (unsigned long) message_id));
 
 		if ( message_id == 0 ) {
 			mx_stack_traceback();
